Made sidebar hader a standalone component
